Rename pkg to versionreader
